Added inputs and secrets for build and publish workflow
This commit introduces new inputs and secrets to the multi-platform build and publish workflow. These additions provide the necessary credentials and configurations for building and publishing applications across different platforms.

Specifically, inputs for publishing on iOS, desktop, and web are added, along with secrets for keystore, GitHub, Google services, Firebase, and Play Store. This change enhances the automation and flexibility of the workflow.