chore: Refactor KMP Platform Build and Publish Workflow
Removed unnecessary input parameters for package names and hardcoded them directly in the workflow. This simplifies the workflow and reduces the number of required inputs.

Removed the 'build_ios' input and made iOS app building a default step in the workflow. This ensures the iOS app is always built alongside other platforms.