chore: Updated dependencies and disabled code signing for development builds
This commit updates various dependencies in the Gemfile.lock, including:

- Updating `aws-partitions` to `1.1028.0`.
- Updating `faraday-multipart` to `1.1.0`.
- Updating `json` to `2.9.1`.

This commit also disables code signing for development builds within the Fastfile, using `skip_codesigning: "true"`. This is presumably to simplify the development workflow.

Finally, it replaces the direct read of the `.ruby-version` file with `File.read(File.join(__dir__, ".ruby-version")).strip` which uses the correct directory when executing the Gemfile.